requirements.txt
requests
beautifulsoup4
lxml
pymongo
python-dotenv
playwright
docling


